Changes to avoid warnings in windows builds. 1. BlockAllocator::parsimony_index_bound is now an int 2. In ParsimonyRouter::constructTree() an intptr_t in a for-loop     is now an int (2 billion nodes in a tree?  unlikely!) 3. In PhyloTree::computeParsimonyTreeNew,  the same. 4. In PhyloTree::optimizePlacementRegion, some static casts     are needed (# of neighbors of a node, to int - safe),     size_t branch index to an int (because Neighbor::id is int)      (# of branches in a tree, int - 2 billion branches in a tree?      unlikely!) (there are two of these).

+++ tree/parsimonyjoining.cpp
@@ -483,7 +483,7 @@ public:
         TypedTaxaToPlace<RoutedTaxon> candidates;
         candidates.resize(nseq);
         int first_new_interior_id = tree.renumberInternalNodes();
-        for (intptr_t i=3; i<nseq; ++i) {
+        for (int i=3; i<nseq; ++i) {
             int         taxonId   = taxon_order[i];
             std::string taxonName = tree.aln->getSeqName(taxonId);
             candidates[i].initialize(&block_allocator, first_new_interior_id+i-3, taxonId, taxonName, false);
