+++ tree/phylotreepars.cpp
@@ -1419,7 +1419,7 @@ int PhyloTree::computeParsimonyTreeNew(const char *out_prefix,
     TypedTaxaToPlace<TaxonToPlace> candidates;
     candidates.resize(count_to_add);
     int first_new_interior_id = renumberInternalNodes();
-    for (intptr_t i=0; i<count_to_add; ++i) {
+    for (int i=0; i<count_to_add; ++i) {
         int         taxonId   = taxon_order[i+leafNum];
         std::string taxonName = aln->getSeqName(taxonId);
         candidates[i].initialize(&block_allocator, first_new_interior_id+i,
